+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-5_6/src/main/java/co/elastic/apm/agent/esrestclient5_6/package-info.java
@@ -23,6 +23,6 @@
  * #L%
  */
 @NonnullApi
-package co.elastic.apm.agent.es.restclient.v5_6;
+package co.elastic.apm.agent.esrestclient5_6;
 
 import co.elastic.apm.agent.sdk.NonnullApi;
