+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-6_4/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,2 +1,2 @@
-co.elastic.apm.agent.es.restclient.v6_4.ElasticsearchClientSyncInstrumentation
-co.elastic.apm.agent.es.restclient.v6_4.ElasticsearchClientAsyncInstrumentation
+co.elastic.apm.agent.esrestclient6_4.ElasticsearchClientSyncInstrumentation
+co.elastic.apm.agent.esrestclient6_4.ElasticsearchClientAsyncInstrumentation
